Add Slice and SlicePtr types
A common use case for package set is to build a deserialization package
that unmarshals data into a type T, appends it to a provided []T, and
returns the result.

Slice and SlicePtr facilitate such deserializers.

SlicePtr ensures a provided interface{} value is a *[]T or pointer chain
to []T that can be manipulated.

Slice allows easier element T creation and appending to the slice.

It's possible Slice and SlicePtr should be combined into a single type.