Fix PreparedMapping.Rebind bug and refactor tests
PreparedMapping.Rebind would clear the internal error even if the error
represented ErrNoPlan.  This means a PreparedMapping could be made, have
an erroneous plan set (i.e. with unknown field), and then a call to
Rebind would clear the ErrNoPlan.  Of course the plan is still invalid
and the error should not have been cleared -- this is fixed.

All BoundMapping and PreparedMapping tests in mapper_combined_test.go
have been rewritten to run in two iterations.  The first iteration
performs Mapper.Bind or Mapper.Prepare to create the Bound or Prepared
mapping and the second iteration uses Rebind before running the tests
again on a new value.  This change correctly covers the bug mentioned
above and should prevent future bugs from Rebind clearing internal state
from appearing in BoundMapping and PreparedMapping types.

Finally the private pkgerr() method on both BoundMapping and
PreparedMapping have been removed, methods in Mapper adjusted
accordingly, and the internal err field of PreparedMapping is always set
to an instance of pkgerr.