Fix performance bottleneck in Mapper.Prepare
During Mapper.Prepare the PreparedMapping.err field was set to an
instance of pkgerr and used string concatenation.  Under load this hurts
performance and memory consumption.

Fixed by setting err field to ErrNoPlan and the methods will perform the
creation of pkgerr.  The performance issue still exists but is now moved
to the error path.