Apply bug fix in PreparedMapping.Plan
The Plan method did not reset the valid flag before iterating the fields
of the plan.  This means a good plan could be applied, followed by a bad
plan with unknown field, and the prepared mapping would have its
internal `valid` flag still set to `true`.

This commit correctly sets valid to false before iterating the fields
and adds an appropriate test.