Add test for TreatAsScalar
While updating my sqlh package to use the next expected version of set
it was discovered TreatAsScalar did not apply correctly with the
internal change in set to use the path subpackage.  When Mapper looks in
the path.Tree for the generated name it should expect TreatAsScalar
types to be in the Branches field of the tree instead of the Leaves.
This is fixed.